refactor(authorization): change auth and validation methods to Author…

+++ src/main/java/com/aps/recycleplace/implementation/services/ListingsService.java
@@ -13,7 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.RequestBody;
-
+import org.springframework.web.bind.annotation.RequestHeader;
 import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
@@ -31,10 +31,10 @@ public class ListingsService {
 		this.orderRepository = orderRepository;
 	}
 
-	public void addNewListings(HttpServletResponse response,
-			@RequestBody Listagem listingObj,
-			@CookieValue(value = "token", defaultValue = "undefined") String authString) {
-		Usuario usuariodb = usuarioRepository.findUsuarioByAuth(authString).orElse(null);
+	public void addNewListings(HttpServletResponse response, @RequestBody Listagem listingObj,
+			String BearerToken) {
+		System.out.println("\nBearerToken\n" + BearerToken + "\n");
+		Usuario usuariodb = usuarioRepository.findUsuarioByAuth(BearerToken.split(" ")[1]).orElse(null);
 		if (usuariodb == null) {
 			response.setStatus(401);
 			return;
@@ -43,46 +43,54 @@ public class ListingsService {
 		listingsRepository.save(listingObj);
 	}
 
-	public void removeListings(HttpServletResponse response,
-			Long listingId,
-			String authString) {
-		try {
-			listingsRepository.deleteById(listingId);
-		} catch (Exception e) {
-			response.setStatus(406);
+	public void removeListings(HttpServletResponse response, Long listingId, String Bearer) {
+		Usuario usuarioDb = usuarioRepository.findUsuarioByAuth(Bearer.split(" ")[1]).orElse(null);
+		Listagem listingDb = listingsRepository.findById(listingId).orElse(null);
+
+		if (usuarioDb != null) {
+			if (usuarioDb.getId() == listingDb.getUsuario().getId()) {
+				try {
+					listingsRepository.deleteById(listingId);
+				} catch (Exception e) {
+					response.setStatus(406);
+				}
+			}
 		}
 	}
 
-	public List<ListingsDTO> getListingsBySearch(HttpServletResponse response,
-			String authString,
-			String searchParam) {
-		List<ListingsDTO> searchResult;
-		Usuario usuarioDb = usuarioRepository.findUsuarioByAuth(authString).orElse(null);
+	public List<Listagem> getAllListings(HttpServletResponse response, String BearerToken) {
+		System.out.println("\nBearerToken\n" + BearerToken + "\n");
+
+		Usuario usuarioDb = usuarioRepository.findUsuarioByAuth(BearerToken.split(" ")[1]).orElse(null);
 		if (usuarioDb == null) {
 			response.setStatus(401);
 			return null;
 		}
 
-		if (searchParam == null) {
-			searchResult = listingsRepository.findByusuario_id(usuarioDb.getId());
-		} else {
-			searchResult = listingsRepository.findBySearchParam(searchParam);
-		}
+		List<Listagem> searchResult;
+
+		searchResult = listingsRepository.findAll();
 
 		if (searchResult.isEmpty()) {
 			response.setStatus(404);
+
 			return null;
 		}
 
 		return searchResult;
 	}
 
-	public void createOrder(HttpServletResponse response,
-			String authString,
-			Long id) {
-		Usuario usuarioDb = usuarioRepository.findUsuarioByAuth(authString)
-				.orElse(null);
-		Listagem listingDb = listingsRepository.findById(id).orElse(null);
+	public void createOrder(HttpServletResponse response, String Bearer, Long listingId) {
+		Usuario usuarioDb = usuarioRepository.findUsuarioByAuth(Bearer.split(" ")[1]).orElse(null);
+		Listagem listingDb = listingsRepository.findById(listingId).orElse(null);
+
+		UsuarioOrdem exists =
+				orderRepository.findToValidate(usuarioDb.getId(), listingDb.getId()).orElse(null);
+
+		if (exists != null) {
+			response.setStatus(409);
+			return;
+		}
 		if (usuarioDb == null || listingDb == null) {
 			System.out.println("-------------------------------------------");
 			System.out.println(usuarioDb);
@@ -96,9 +104,8 @@ public class ListingsService {
 		orderRepository.save(usuarioOrdemObj);
 	}
 
-	public List<OrderDTO> getMyOrders(HttpServletResponse response,
-			String authString) {
-		Usuario usuarioDb = usuarioRepository.findUsuarioByAuth(authString).orElse(null);
+	public List<OrderDTO> getMyOrders(HttpServletResponse response, String Bearer) {
+		Usuario usuarioDb = usuarioRepository.findUsuarioByAuth(Bearer.split(" ")[1]).orElse(null);
 		if (usuarioDb == null) {
 			response.setStatus(401);
 			return null;
@@ -114,10 +121,8 @@ public class ListingsService {
 
 	}
 
-	public void removeOrder(HttpServletResponse response,
-			String authString,
-			Long orderParam) {
-		Usuario usuarioDb = usuarioRepository.findUsuarioByAuth(authString).orElse(null);
+	public void removeOrder(HttpServletResponse response, String Bearer, Long orderParam) {
+		Usuario usuarioDb = usuarioRepository.findUsuarioByAuth(Bearer.split(" ")[1]).orElse(null);
 		UsuarioOrdem ordemDb = orderRepository.findOrderByOwnerId(usuarioDb.getId()).orElse(null);
 
 		if (usuarioDb == null || ordemDb == null) {
